app/requirements.txt
web3==5.28.0
eth-tester==0.6.0b6
py-solc-x==1.1.1
flask==2.0.2
Flask-SQLAlchemy==2.5.1
psycopg2==2.9.3
flask-restful==0.3.9
marshmallow==3.14.1
pyjwt==2.3.0
python-binance==1.0.15
celery==5.2.3